requirements.txt
requests
flask
flask-sqlalchemy
mysqlclient
flask_script